Add segment name to error message
Add the name of the memory segment to the "Leader should be started
after Follower" error message.  This makes it easier to see where the
error is coming from in a setup with different Leaders/Followers.

Use the fmt library (new dependency) for more nicely inserting the name
into the string.